+++ ctf/commands/init.py
@@ -1,12 +1,12 @@
-import os
+import importlib.resources
 import shutil
+from pathlib import Path
 
 import typer
 from typing_extensions import Annotated
 
 from ctf import ENV
-from ctf.logger import LOG
-from ctf.utils import get_ctf_script_templates_directory
+from ctf.common.logger import LOG
 
 app = typer.Typer()
 
@@ -16,8 +16,8 @@ app = typer.Typer()
 )
 def init(
     path: Annotated[
-        str, typer.Argument(help="Directory in which to initialize a CTF")
-    ] = "",
+        Path | None, typer.Argument(help="Directory in which to initialize a CTF")
+    ] = None,
     force: Annotated[
         bool,
         typer.Option(
@@ -27,40 +27,36 @@ def init(
 ) -> None:
     # If path is not set, take the one from --location or CTF_ROOT_DIR, else it's the current directory.
     if not path:
-        path = (
-            str(ENV.get("CTF_ROOT_DIR"))
-            if "CTF_ROOT_DIR" in ENV
-            else os.path.join(os.getcwd(), ".")
-        )
+        path = Path(ENV.get("CTF_ROOT_DIR", "."))
+
+    path = path.expanduser().resolve()
 
     created_directory = False
-    created_assets: list[str] = []
+    created_assets: list[Path] = []
     try:
-        if not os.path.isdir(path):
-            os.mkdir(path)
+        if not path.is_dir():
             LOG.info(f'Creating directory "{path}"')
+            path.mkdir()
             created_directory = True
         elif (
-            os.path.isdir(os.path.join(path, "challenges"))
-            or os.path.isdir(os.path.join(path, ".deploy"))
+            (path / "challenges").is_dir() or (path / ".deploy").is_dir()
         ) and not force:
             LOG.error(
                 f'Directory "{path}" is already initialized. Use --force to overwrite.'
             )
-            exit(code=1)
+            exit(1)
 
-        for asset in os.listdir(
-            p := os.path.join(get_ctf_script_templates_directory(), "init")
-        ):
-            dst_asset = os.path.join(path, asset)
-            if os.path.isdir(src_asset := os.path.join(p, asset)):
-                shutil.copytree(src_asset, dst_asset, dirs_exist_ok=True)
-                LOG.info(f'Created "{dst_asset}" folder')
-            else:
-                shutil.copy(src_asset, dst_asset)
-                LOG.info(f'Created "{dst_asset}" file')
+        with importlib.resources.path("ctf.templates", "init") as templates_location:
+            for asset in templates_location.iterdir():
+                dst_asset: Path = path / asset.name
+                if asset.is_dir():
+                    shutil.copytree(asset, dst_asset, dirs_exist_ok=True)
+                    LOG.info(f'Created "{dst_asset}" folder')
+                else:
+                    shutil.copy(asset, dst_asset)
+                    LOG.info(f'Created "{dst_asset}" file')
 
-            created_assets.append(dst_asset)
+                created_assets.append(dst_asset)
 
     except Exception:
         import traceback
@@ -70,11 +66,11 @@ def init(
             LOG.info(f'Removed created "{path}" folder')
         else:
             for asset in created_assets:
-                if os.path.isdir(asset):
+                if asset.is_dir():
                     shutil.rmtree(asset)
                     LOG.info(f'Removed created "{asset}" folder')
                 else:
-                    os.unlink(asset)
+                    asset.unlink()
                     LOG.info(f'Removed created "{asset}" file')
 
         LOG.critical(traceback.format_exc())